requirements.txt
Flask==3.0.3
Flask_Cors==5.0.0
protobuf==5.28.3
python-dotenv==1.0.1
